fixup! Move master related code out of plugins
